refactor: rename signup module to households
penny/signup.py had outgrown its name: every function in it is
household lifecycle (provisioning, invites, members, rename), and
"household" is the codebase's established term (household_id tenancy,
REQUIREMENTS P10, /api/household routes). It becomes
penny/households.py, and penny/api/signup_routes.py becomes
penny/api/household_routes.py to match billing_routes. Purely
mechanical: imports, docstrings, and the ruff B008 per-file exemption;
no function renames, no route-path changes. Historical docs/plans keep
their old references per the non-canonical artifacts rule.
